enh: convert catalog money fields to Decimal (MONEY-3) (#18)
Convert ProductVariant.current_cogs/base_price and
ProductVariantMarketplace.selling_price/discounted_price from
BigIntegerField to DecimalField(18,2); Shopee identifier fields
untouched. Move price validation in catalog views into serializers,
and round Shopee-outbound prices via round_money_to_int instead of
float, keeping money values off both float and view-layer validation.

+++ apps/catalog/services/product_service.py
@@ -1,4 +1,5 @@
 import logging
+from decimal import Decimal
 from typing import Any
 
 from django.db import transaction
@@ -255,7 +256,9 @@ class ProductService:
                     exc_info=True,
                 )
 
-    def update_variant_base_price(self, variant_id: str, base_price: int) -> dict[str, str | int]:
+    def update_variant_base_price(
+        self, variant_id: str, base_price: Decimal
+    ) -> dict[str, str | Decimal]:
         from apps.catalog.models import ProductVariant
 
         variant = ProductVariant.objects.get(id=variant_id)
